backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic[email]
python-dotenv
pandas
openpyxl
